+++ src/exactly_lib/impls/types/string_matcher/parse/equality.py
@@ -1,80 +1,56 @@
 from typing import Sequence
 
-from exactly_lib.common.help.syntax_contents_structure import SyntaxElementDescription
-from exactly_lib.definitions import formatting, syntax_descriptions
-from exactly_lib.definitions.argument_rendering.path_syntax import the_path_of
+from exactly_lib.definitions import formatting
 from exactly_lib.definitions.cross_ref.app_cross_ref import SeeAlsoTarget
 from exactly_lib.definitions.entity import syntax_elements
 from exactly_lib.impls.types.expression import grammar
-from exactly_lib.impls.types.path import path_relativities, relative_path_options_documentation
+from exactly_lib.impls.types.path import path_relativities
 from exactly_lib.impls.types.string_matcher.impl import equality
-from exactly_lib.impls.types.string_or_path import parse_string_or_path
-from exactly_lib.impls.types.string_or_path.doc import StringOrHereDocOrFile
+from exactly_lib.impls.types.string_source import parse as parse_str_src
+from exactly_lib.section_document.element_parsers.ps_or_tp.parser import PARSE_RESULT
+from exactly_lib.section_document.element_parsers.ps_or_tp.parsers import ParserFromTokenParserBase
 from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser
+from exactly_lib.tcfs.path_relativity import RelOptionType
 from exactly_lib.type_val_deps.types.string_matcher import StringMatcherSdv
 from exactly_lib.util.cli_syntax.elements import argument as a
 from exactly_lib.util.textformat.structure.core import ParagraphItem
 from exactly_lib.util.textformat.textformat_parser import TextParser
 
-_EXPECTED_SYNTAX_ELEMENT_FOR_EQUALS = 'EXPECTED'
-
-EXPECTED_FILE_REL_OPT_ARG_CONFIG = parse_string_or_path.CONFIGURATION
+REL_OPT_CONFIG = path_relativities.all_rel_options_config(RelOptionType.REL_HDS_CASE)
 
 
-def parse(token_parser: TokenParser) -> StringMatcherSdv:
-    token_parser.require_has_valid_head_token(_EXPECTED_SYNTAX_ELEMENT_FOR_EQUALS)
-    expected_contents = parse_string_or_path.parse_from_token_parser(
-        token_parser,
-        EXPECTED_FILE_REL_OPT_ARG_CONFIG,
-        consume_last_here_doc_line=False)
+class EqualsParser(ParserFromTokenParserBase[StringMatcherSdv]):
+    def __init__(self):
+        super().__init__(False, False)
+        self._str_src_parser = parse_str_src.string_source_parser(REL_OPT_CONFIG)
 
-    return equality.sdv(expected_contents)
+    def parse_from_token_parser(self, parser: TokenParser) -> PARSE_RESULT:
+        expected_contents = self._str_src_parser.parse_from_token_parser(parser)
+
+        return equality.sdv(expected_contents)
 
 
 class Description(grammar.PrimitiveDescriptionWithNameAsInitialSyntaxToken):
-    def __init__(self):
-        self._string_or_here_doc_or_file_arg = StringOrHereDocOrFile(
-            _EXPECTED_PATH_NAME,
-            _RELATIVITY_OF_EXPECTED_PATH_NAME,
-            path_relativities.all_rel_options_arg_config(_EXPECTED_PATH_NAME),
-            the_path_of('the file that contains the expected contents.')
-        )
-
     @property
     def argument_usage_list(self) -> Sequence[a.ArgumentUsage]:
-        return self._string_or_here_doc_or_file_arg.argument_usage(
-            a.Multiplicity.MANDATORY),
+        return [
+            syntax_elements.STRING_SOURCE_SYNTAX_ELEMENT.single_mandatory
+        ]
 
     @property
     def description_rest(self) -> Sequence[ParagraphItem]:
         tp = TextParser({
-            'HERE_DOCUMENT': formatting.syntax_element_(syntax_elements.HERE_DOCUMENT_SYNTAX_ELEMENT),
-            'Sym_refs_are_not_substituted': syntax_descriptions.symbols_are_not_substituted_in(
-                'the file ' + self._string_or_here_doc_or_file_arg.path_name
-            ),
+            'STRING_SOURCE': formatting.syntax_element_(syntax_elements.STRING_SOURCE_SYNTAX_ELEMENT),
         })
         return tp.fnap(_DESCRIPTION)
 
     @property
-    def syntax_elements(self) -> Sequence[SyntaxElementDescription]:
+    def see_also_targets(self) -> Sequence[SeeAlsoTarget]:
         return [
-            relative_path_options_documentation.path_element_2(
-                self._string_or_here_doc_or_file_arg.path_argument_configuration
-            )
+            syntax_elements.STRING_SOURCE_SYNTAX_ELEMENT.cross_reference_target
         ]
 
-    @property
-    def see_also_targets(self) -> Sequence[SeeAlsoTarget]:
-        return self._string_or_here_doc_or_file_arg.see_also_targets()
-
-
-_EXPECTED_PATH_NAME = 'PATH-OF-EXPECTED'
-_RELATIVITY_OF_EXPECTED_PATH_NAME = 'RELATIVITY-OF-EXPECTED-PATH'
 
 _DESCRIPTION = """\
-Matches iff the string is equal to a given
-string, {HERE_DOCUMENT} or contents of a file.
-
-
-{Sym_refs_are_not_substituted}
+Matches iff the string is equal to {STRING_SOURCE}.
 """