+++ test/exactly_lib_test/impls/types/string_matcher/equals.py
@@ -4,7 +4,7 @@ from abc import ABC
 from typing import Sequence
 
 from exactly_lib.impls.types.string_matcher.parse.equality import \
-    EXPECTED_FILE_REL_OPT_ARG_CONFIG
+    REL_OPT_CONFIG
 from exactly_lib.tcfs.path_relativity import RelOptionType
 from exactly_lib.test_case.app_env import ApplicationEnvironment
 from exactly_lib.test_case.path_resolving_env import PathResolvingEnvironmentPreOrPostSds
@@ -84,10 +84,10 @@ _RELATIVITY_OPTION_CONFIGURATIONS_FOR_EXPECTED_FILE = [
     rel_opt.default_conf_rel_any(RelOptionType.REL_HDS_CASE),
     rel_opt.symbol_conf_rel_any(RelOptionType.REL_TMP,
                                 'EXPECTED_FILE_SYMBOL',
-                                EXPECTED_FILE_REL_OPT_ARG_CONFIG.options.accepted_relativity_variants),
+                                REL_OPT_CONFIG.accepted_relativity_variants),
     rel_opt.symbol_conf_rel_any(RelOptionType.REL_HDS_CASE,
                                 'EXPECTED_FILE_SYMBOL',
-                                EXPECTED_FILE_REL_OPT_ARG_CONFIG.options.accepted_relativity_variants),
+                                REL_OPT_CONFIG.accepted_relativity_variants),
 ]
 
 _TRANSFORMER_SYMBOL_NAME = 'the_transformer'